--- modules/dkan_resourcemap_entity/dkan_resourcemap_entity.links.contextual.yml
@@ -1,10 +1,0 @@
-entity.resourcemap.edit_form:
-  route_name: entity.resourcemap.edit_form
-  group: resourcemap
-  title: Edit
-
-entity.resourcemap.delete_form:
-  route_name: entity.resourcemap.delete_form
-  group: resourcemap
-  title: Delete
-  weight: 10
