--- modules/dkan_resourcemap_entity/dkan_resourcemap_entity.module
@@ -1,38 +1,0 @@
-<?php
-
-/**
- * @file
- * Provides a resourcemap entity type.
- */
-
-use Drupal\Core\Render\Element;
-
-/**
- * Implements hook_theme().
- */
-function dkan_resourcemap_entity_theme() {
-  return [
-    'resourcemap' => [
-      'render element' => 'elements',
-    ],
-  ];
-}
-
-/**
- * Prepares variables for resourcemap templates.
- *
- * Default template: resourcemap.html.twig.
- *
- * @param array $variables
- *   An associative array containing:
- *   - elements: An associative array containing the resourcemap information and any
- *     fields attached to the entity.
- *   - attributes: HTML attributes for the containing element.
- */
-function template_preprocess_resourcemap(array &$variables) {
-  $variables['view_mode'] = $variables['elements']['#view_mode'];
-  foreach (Element::children($variables['elements']) as $key) {
-    $variables['content'][$key] = $variables['elements'][$key];
-  }
-}
-
