--- modules/dkan_resourcemap_entity/dkan_resourcemap_entity.permissions.yml
@@ -1,3 +1,0 @@
-administer resourcemap:
-  title: 'Administer resourcemap settings'
-  restrict access: true
