--- modules/dkan_resourcemap_entity/src/ResourcemapInterface.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Drupal\dkan_resourcemap_entity;
-
-use Drupal\Core\Entity\ContentEntityInterface;
-
-/**
- * Provides an interface defining a resourcemap entity type.
- */
-interface ResourcemapInterface extends ContentEntityInterface {
-
-}
