+++ modules/metastore/src/ResourceMappingInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Drupal\metastore;
+
+use Drupal\Core\Entity\ContentEntityInterface;
+
+/**
+ * Provides an interface defining a resource mapping entity type.
+ */
+interface ResourceMappingInterface extends ContentEntityInterface {
+
+}
